fix: iOS hot reload dropped in native-direct boots
Hot reload triggers (and OTA reload requests) are delivered via
reloadWebViewNotification, but its only observer was registered inside
WebView.makeUIView. A native-direct boot withholds the WebView entirely —
and ContentView unmounts it whenever native UI is active — so fully-native
apps never registered an observer and every reload trigger was silently
dropped: the watcher synced files and connected to port 9999, but the app
never rebooted PHP.

Move the reload logic off the WebView Coordinator into a new
HotReloadCoordinator singleton that registers for the notification at app
launch, independent of the WebView lifecycle. The WebView, when one
exists, is resolved lazily through SharedWebView.shared. Registration is
unconditional (not DEBUG-only) because AppUpdateManager posts the same
notification after applying an OTA update in production.

Android is unaffected — its watcher thread starts in MainActivity.onCreate
independent of the WebView and already branches on NativeUIBridge.isActive.
